IterativeScoring: always update matrices
There were a few places which avoided work, relying on an assumption
that the matrices are still in their initial state (e.g. all zeroes).
This might not be true for an object that's been in use for a while, so
I changed them to explicitly update stuff just to be safe.